more changes

+++ clientsrc/src/components/Note.vue
@@ -1,26 +1,23 @@
 
 <template>
-<div>
-
   <div>
-    <button type="button" class="close text-danger" @click="deleteNote()">
-      <span >&times;</span>
-    </button>
-    <p>Body: {{noteData.content}}</p>
-    <p>Email: {{noteData.creatorEmail}}</p>
+    <div>
+      <button v-if="this.$auth.userInfo.email == this.noteData.creatorEmail" type="button" class="close text-danger" @click="deleteNote()">
+        <span>&times;</span>
+      </button>
+      <p>Body: {{noteData.content}}</p>
+      <p>Email: {{noteData.creatorEmail}}</p>
 
-        <!-- <form @submit.prevent="changeNote()">
+      <!-- <form @submit.prevent="changeNote()">
           <div class="form-group">
             <label for="body">Note</label>
             <input type="text" name="body" id="" class="form-control" placeholder="enter Note..." aria-describedby="helpId" v-model="noteData.content" required>
           </div>
 
           <button type="submit" class="btn btn-success">Save Note</button>
-        </form> -->
+      </form>-->
+    </div>
   </div>
-
-
-</div>
 </template>
 
 <script>
@@ -32,14 +29,16 @@ export default {
   },
   methods: {
     deleteNote() {
-      console.log("deleteNote", this.noteData)
-      this.$store.dispatch("deleteNote", this.noteData);
+      if (window.confirm("Are you sure?")) {
+        //console.log("deleteNote", this.noteData);
+        this.$store.dispatch("deleteNote", this.noteData);
+      }
     },
     changeComment() {
       //console.log("changeBlog", this.blogDetails)
       //this.$router.push({ name: "Home" });
       this.$store.dispatch("changeNote", this.noteData);
-    },
+    }
   }
 };
 </script>
